== Guild Window & more == *Please don't forget to add the upgrade_15-3athena_svn467.sql file!
*Some misc cleaning.
-Moved some stable costume items.
-Adjusted a couple skills that have infinite timers to not display negative time to the client.

*Fixed Chorus skills not checking for additional partners.

*Refactored clif functions for unit names.
-Renamed clif_charnameack to clif_name and added additional parameters to follow other clif functions.
-Removed clif_charnameupdate completely since it was just a copy of the BL_PC code in clif_charnameack and replaced all calls with clif_name_area.

*Fixed a potential map-server crash.

*Added some checks to guild leader change
-Introduces two new configurations:
--Allow guild leader changes in WoE (yes/no)
---Default: no
--Delay between the changes (in minutes)

*Fixed ZC_ACK_REQNAME_BYGID packet structure for 2018+ clients.
-This also fixed the guild window.
-finally fixes issue #15

+++ src/map/pc.h
@@ -267,7 +267,7 @@ struct map_session_data {
 	short skillid_dance,skilllv_dance;
 	short cook_mastery; // range: [0,1999] [Inkfish]
 	struct skill_cooldown_entry *scd[MAX_SKILLCOOLDOWN]; // Skill Cooldown
-	int cloneskill_idx, ///Stores index of copied skill by Intimidate/Plagiarism
+	short cloneskill_idx, ///Stores index of copied skill by Intimidate/Plagiarism
 		reproduceskill_idx; ///Stores index of copied skill by Reproduce
 	int menuskill_id, menuskill_val, menuskill_val2, menuskill_itemused;
